Crate iris_compiler

Macros for defining subscriptions in Iris.

Every Iris subscription requires a callback, a filter, and one or more data types. Many simple applications can get by with just the callback macro, leveraging Iris’ filter DSL (which can filter on protocols and protocol fields) and the sample data types in the iris_datatypes crate.

For more complex use-cases, developers can define custom filters with filter and data types with [fn@datatypes]. Stateful callbacks, filters, and data types — i.e., those that accumulate data over the lifetime of a connection – use the callback_fn, filter_fn, and datatype_fn, respectively, to annotate struct methods.

Macro Arguments

Because Rust macros cannot access state beyond their immediate input, all *_fn annotations must take in the name of the struct they are associated with as their first argument. The #callback macro must take in its filter pattern as its first argument.

Other optional arguments:

level: Explicitly indicate the state transition that a function should be invoked at. &-separated.

parsers: Explicitly specify session-level protocol parsers, &-separated (e.g., “http&tls”).

tracked: Data type only; see datatype.

Notes on parsers:

Iris will infer and register protocol parsers based on data types and filters; for example, if a function requests a TLS handshake or filters for “tls”, then Iris will register the “tls” parser.

Developers should use this to explicitly register additional parsers. For example, the “Session”, “StateTxData”, and “SessionProto” types do not register any parsers.

Return values

Streaming callback functions must return a boolean value, which can be “false” to unsubscribe from a connection (i.e., stop receiving updates for that connection). Note that if one function in a callback group (i.e., a struct method) returns false, the entire callback (all struct methods) is considered “unsubscribed”.

Filter functions must return an [iris_core::subscription::FilterResult].

Filter DSL

Iris’ filter DSL implements Retina’s filter language with added support for a contains operator and raw byte matching (see below). Iris developers can also filter on custom predicates by referring to them by name (function name for filter functions or struct name for stateful filters).

Added Binary Comparison Operators

Operator	Alias	Description	Example

~b		Byte regular expression match	ssh.protocol_version_ctos ~b '(?-u)^\x32\\.\x30$'

contains		Check if right appears in left	ssh.key_exchange_cookie_stoc contains |15 A1|

not contains	!contains	Check that right doesn’t appear in left	ssh.key_exchange_cookie_stoc not contains |15 A1|

Added Field Types (RHS values)

| Type | Example | | Byte | \|32 2E 30\| |

Attribute Macros

cache_file

If a crate or target needs to export data types, filters, or callbacks to another crate or target, it must specify a file to cache parsed data in. Other crates or targets access these exported types using input_files. Rust macros can’t maintain local state across targets, and this is a simple workaround. See the iris_datatypes crate source for an example of using cache_file, and see the examples directory for applications that use input_files to import definitions from other crates or targets.

callback

Indicate that a struct or function is a callback. Must specify a filter as a string.

callback_fn

Indicate that a struct method should be invoked by Iris as part of constructing a stateful callback

datatype

Indicate that a struct or type is an Iris data type.

datatype_fn

Indicate that a struct method should be invoked by Iris as part of constructing a data type. These functions can take in any “built-in” data types[fn@BUILTIN_TYPES] (except FilterStr, which applies only to callbacks): L4Pdu, StateTxData, StateTransition, StateTxData, Session, SessionProto, CoreId, or StateTransition.

filter

Indicate that a struct or function is a filter.

filter_fn

Indicate that a struct method should be invoked by Iris as part of constructing a stateful filter.

input_files

See cache_file.

iris_end_macros

In practice, Rust processes procedural macros in the order they appear in imports and in files (though technically this behavior is not guaranteed). Iris use iris_end_macros to infer that it has read all macro inputs and can begin generating code. This must go in the main file of every binary target. Generally, putting it on the main function makes sense.