iris_compiler

Attribute Macro datatype

#[datatype]

Indicate that a struct or type is an Iris data type.

Example usage:

#[datatype("L7EndHdrs,parsers=dns")]
pub type DnsTransaction = Box<Dns>;

Developers can explicitly tag a datatype as “tracked”. This indicates that updating the data type is computationally- and memory-intensive. The data type must implement the “Tracked” trait, and Iris will “clear” and stop invoking APIs on the data type if all connections requiring it go out of scope. Requiring Iris to track a data type adds some overhead and limits Iris’s ability to optimize filter trees.